Fixes #34436 - dont delete shared pulp3 repos for composites

+++ app/lib/actions/pulp3/content_view/delete_repository_references.rb
@@ -11,14 +11,24 @@ module Actions
         def invoke_external_task
           tasks = []
           content_view = ::Katello::ContentView.find(input[:content_view_id])
-          content_view.repository_references.each do |repository_reference|
+          to_delete = content_view.repository_references.select do |repository_reference|
             repo = repository_reference.root_repository.library_instance
-            #force pulp3 in case we've done migrations, but haven't switched over yet
-            tasks << repo.backend_service(smart_proxy, true).delete_repository(repository_reference)
+            if delete_href?(repository_reference.repository_href, content_view)
+              #force pulp3 in case we've done migrations, but haven't switched over yet
+              tasks << repo.backend_service(smart_proxy, true).delete_repository(repository_reference)
+              true
+            else
+              false
+            end
           end
-          content_view.repository_references.destroy_all
+          to_delete.each(&:destroy)
 
           output[:pulp_tasks] = tasks
+        end
+
+        #migrated composites may have the same RepositoryReference as their component
+        def delete_href?(href, content_view)
+          ::Katello::Pulp3::RepositoryReference.where(:repository_href => href).where.not(:content_view_id => content_view.id).empty?
         end
       end
     end
